Checks for existing parsed copy of any tune played
Using the extracted Title from the ABC tune, together with the integer
representing the instrument used, now we check with readProjectData to
see if we already have it.  This makes your parsed tunes accompany your
project, so it is a little bigger but a lot faster.  These data in the
Data.plist file would go along with your project if you were App
Storing it…